chore: refactor code and add test
